crawler/requirements.txt
newspaper3k
pymongo
requests
pyyaml
beautifulsoup4
lxml
